requirements.txt
Django
djangorestframework
pytest
pytest-django
requests
boto3
django-cors-headers